+++ RayTraceRenderer/color.h
@@ -6,15 +6,15 @@ class Color : public Tuple3d
 public:
 	//constructors
 	Color() : Color(0, 0, 0) {};
-	Color(float, float, float);
+	Color(double, double, double);
 	//~Color();
 
 	//getters and setters
-	void setRed(float);
-	float getRed();
-	void setGreen(float);
-	float getGreen();
-	void setBlue(float);
-	float getBlue();
+	void setRed(double);
+	double getRed();
+	void setGreen(double);
+	double getGreen();
+	void setBlue(double);
+	double getBlue();
 };
