+++ RayTraceRenderer/tuple_features.cpp
@@ -1,9 +1,9 @@
 #include "pch.h"
 
-Tuple3d point(float x, float y, float z) {
+Tuple3d point(double x, double y, double z) {
 	return Tuple3d(x, y, z, 1.0);
 }
 
-Tuple3d vector(float x, float y, float z) {
+Tuple3d vector(double x, double y, double z) {
 	return Tuple3d(x, y, z, 0.0);
 }
