+++ RayTraceRenderer/tuple_features.h
@@ -1,2 +1,2 @@
-Tuple3d point(float, float, float);
-Tuple3d vector(float, float, float);
+Tuple3d point(double, double, double);
+Tuple3d vector(double, double, double);
